Add files via upload
The business goal of this project was to model the prices of houses whith the available independent variable. This model will further be used by the management to understand how exactly the prices vary with the variables. They can accordinly manipulate the strategy of the firm and concentrate on areas that will yield high returns. Further, the model will be a good way for the management to understand the pricing dynamics of a new market.